refactor: move useKeyboard from shared to app-level hooks
The useKeyboard hook imports from features/grid-editor, which violates the
architectural principle that shared code should not depend on feature code.

Moving it to src/hooks/ (app-level) is the correct fix because:
- App-level hooks can depend on features, shared, and core
- useKeyboard is app-specific (not reusable across different apps)
- This eliminates the shared → feature layer boundary violation

The import hierarchy is now properly maintained:
  features → shared → core (shared does not import from features)
